Correct the frame interval time setting

+++ Modbus/Modbus/ModbusRTUSlave/MDS_RTU_Fun.h
@@ -43,7 +43,7 @@ typedef struct{
 	uint32 lastSendTimes;
 	
 	/*Frame interval time 3.5T*/
-	uint16 frameIntervalTime;
+	float frameIntervalTime;
 	
 	/*CRC16 received*/
 	uint16 CRC16Update;
